+++ .idea/compiler.xml
@@ -8,6 +8,7 @@
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
         <module name="CustomTicketHub" />
+        <module name="TicketHub" />
       </profile>
     </annotationProcessing>
   </component>
