+++ src/main/java/io/github/arkery/customtickethub/Enums/Priority.java
@@ -1,0 +1,8 @@
+package io.github.arkery.customtickethub.Enums;
+
+public enum Priority {
+
+    LOW, MEDIUM, HIGH, CRITICAL;
+
+
+}
