+++ src/main/java/io/github/arkery/customtickethub/Enums/Status.java
@@ -1,0 +1,6 @@
+package io.github.arkery.customtickethub.Enums;
+
+public enum Status {
+
+    OPENED, INPROGRESS, RESOLVED, CLOSED
+}
