+++ src/main/java/io/github/arkery/customtickethub/TicketHub.java
@@ -2,7 +2,7 @@ package io.github.arkery.customtickethub;
 
 import org.bukkit.plugin.java.JavaPlugin;
 
-public class CustomTicketHub extends JavaPlugin {
+public class TicketHub extends JavaPlugin {
 
     @Override
     public void onDisable(){
