+++ src/main/java/io/github/arkery/customtickethub/TicketSystem/Core.java
@@ -1,0 +1,8 @@
+package io.github.arkery.customtickethub.TicketSystem;
+
+public class Core {
+
+
+
+    
+}
