+++ src/plugin.yml
@@ -1,7 +1,7 @@
 name: Custom_Ticket_Hub
 version: 1.0
 author: arkery
-main: java.io.github.arkery.customtickethub.CustomTicketHub
+main: java.io.github.arkery.customtickethub.TicketHub
 api-version: 1.14
